splitted LayerSpec to few files
